fix kd_fof mindist bug.
As the referee pointed out mindist and bound mindist are different.
We do not know of the true mindist thus cannot directly apply
the single edge self-connected optimization. What we can do is
to terminate the iteration as soon as a first pair is detected; this
on average is not much worse than the old solution.

We also add a skip_symmetric mode to kd_enum_check. This will
skip the distance computation of half of the pairs, including skipping
the self pairs. these changes need to be rebased to master.

We will redo the tests based on master.