Change definition of tree->ind
ind is the initial 'selection' of entries from the input data as absolute
indices, and
to be initialized by the caller. This allows using a subset of data
without actually chopping off the input data.

The returned indices from kdcount is always the absolute indices.

Next commit will add this to the python api.